Split api layer from file reader (#365)
The update makes updates to the SourceRecordIterator to remove the
requirement for a S3Client and specific S3 knowledge from the iterator.

The iterator will now also call for more files after the initial set of
files has been processed.

The only remaining work to be done is to remove the construction of the
S3Object into an iterator from the SourceRecordIterator in a follow up
PR which will allow it to be completely re-useable.

---------

Signed-off-by: dependabot[bot] <support@github.com>
Signed-off-by: Aindriu Lavelle <aindriu.lavelle@aiven.io>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>