Allow users to vary the size of the byteArrayBuffer in source connectors (#386)
This allows users to decide how large they wish to chunk the byte stream
in the source connectors.
One change was to use the SourceCommonConfig instead of the
abstractConfig in the transformers so that the new calls to get the
maxbytebuffer would be available.

It also adds the stream length to the Transformer which should lead to
improvements specifically in the ByteArrayTransformer.

---------

Signed-off-by: Aindriu Lavelle <aindriu.lavelle@aiven.io>